led driver is working

+++ Pierre/nRF5_SDK_15.0.0_a53641a/integration/nrfx/legacy/nrf_drv_spi.h
@@ -362,7 +362,7 @@ void nrf_drv_spi_uninit(nrf_drv_spi_t const * const p_instance);
 __STATIC_INLINE
 ret_code_t nrf_drv_spi_transfer(nrf_drv_spi_t const * const p_instance,
                                 uint8_t const * p_tx_buffer,
-                                uint8_t         tx_buffer_length,
+                                uint16_t         tx_buffer_length,
                                 uint8_t       * p_rx_buffer,
                                 uint8_t         rx_buffer_length);
 
@@ -490,7 +490,7 @@ void nrf_drv_spi_uninit(nrf_drv_spi_t const * p_instance)
 __STATIC_INLINE
 ret_code_t nrf_drv_spi_transfer(nrf_drv_spi_t const * const p_instance,
                                 uint8_t const * p_tx_buffer,
-                                uint8_t         tx_buffer_length,
+                                uint16_t         tx_buffer_length,
                                 uint8_t       * p_rx_buffer,
                                 uint8_t         rx_buffer_length)
 {
